refactor(frontend): refactor code v2
refactor make-post code
