+++ frontend/src/app/(account)/login/loginCard.tsx
@@ -26,7 +26,7 @@ function LoginCard() {
             src="/images/static/text-icon.svg"
             width={150}
             height={150}
-            alt="instagram"
+            alt="instagram text logo"
           />
         </CardHeader>
